requirements.txt
zipline
trading-calendars
pandas
numpy
pytz
python-binance
